backend/requirements.txt
Flask==3.0.3
Flask-RESTX==1.3.0
Flask-JWT-Extended==4.6.0
Flask-CORS==5.0.0
Flask-Limiter==3.8.0
APScheduler==3.10.4
httpx==0.27.2
pymongo==4.8.0
python-dotenv==1.0.1
tenacity==9.0.0
Werkzeug==3.0.4
dnspython==2.7.0
python-dateutil==2.9.0
pydantic==2.9.2
